pagination list user

+++ src/initData/users/InitUser.js
@@ -1,4 +1,4 @@
-// run node src/init/users/InitUser.js or cd folder and node InitUser.js to create data for userinfos collection
+// run node src/initData/users/InitUser.js or cd folder and node InitUser.js to create data for userinfos collection
 const fs = require('fs');
 require('../../config/Config').connectDB.mongoDB;
 const UserService = require('../../services/UserService');
@@ -17,16 +17,19 @@ exports.initAdmin = async() => {
     const findAdmin = await UserService.findUserByEmail(info.email);
     if (findAdmin) {
         console.log('Admin existed already.');
+        return 0;
     }
     let createAdmin;
     let createProfileAdmin;
     try {
         createAdmin = await UserService.createUser(info);
     } catch (error) {
-        console.log(error)
+        console.log(error);
+        return 0;
     }
     if (!createAdmin) {
         console.log('Can not create admin.');
+        return 0;
     } else {
         info.firstName = 'Admin';
         info.lastName = 'Student';
@@ -38,6 +41,7 @@ exports.initAdmin = async() => {
         } else {
             await UserService.deleteUser(createAdmin._id);
             console.log('Create admin failed.');
+            return 0;
         }
     }
 }
@@ -59,15 +63,21 @@ try {
                 obj.status = statusUser.ACTIVE;
                 obj.statusLogin = false;
                 obj.role = roles.STUDENT;
+                const findUser = await UserService.findUserByEmail(obj.email);
+                if (findUser) {
+                    console.log(`User ${obj.email} existed already.`);
+                    continue;
+                }
                 const userInfo = await UserService.createUser(obj);
                 if (userInfo) {
                     obj.userId = userInfo._id;
                     const createUserProfile = await ProfileService.createProfile(obj);
                     if (createUserProfile) {
-                        console.log(`Create user ${obj.email} success.`)
+                        console.log(`Create user ${obj.email} success.`);
                     } else {
                         await UserService.deleteUser(userInfo._id);
-                        console.log(`Create user ${obj.email} failed.`)
+                        console.log(`Create user ${obj.email} failed.`);
+                        continue;
                     }
                 }
             } catch (error) {
